lint: Detect import cycles, with the resolver the plugin needed
Adds eslint-import-resolver-typescript plus the import/parsers setting
so eslint-plugin-import can resolve .js-suffixed TypeScript imports and
parse the imported files — without both, its graph-based rules see
nothing and pass silently. import/no-cycle is now on (verified against
a deliberate cycle) and immediately caught a real one: config-store and
migrate imported each other, broken by extracting the shared pure
transforms into config/values.ts.

Two rules stay off, documented in place: import/extensions demands .ts
extensions once the resolver maps imports to their .ts files (nodenext
already fails the build on a bad extension), and
import/no-relative-parent-imports turns out to ban depending on parent
directories however the import is spelled — alias imports included —
which is not the specifier rule this repo wants; the core
no-restricted-imports pattern keeps doing that job.
