+++ src/lib/config/config-store.ts
@@ -4,11 +4,10 @@ import Configstore from 'configstore'
 import envPaths from 'env-paths'
 
 import { migrateConfigStore } from './migrate.js'
+import { isStateKey, mergeConfig, splitConfig } from './values.js'
 
 const configFileName = 'cli.json'
 const legacyConfigStoreId = 'seam-cli'
-const currentWorkspaceIdKey = 'current_workspace_id'
-const patKey = 'pat'
 const paths = envPaths('seam', { suffix: '' })
 
 /**
@@ -61,57 +60,6 @@ const createConfigStore = (): PersistentConfigStore => {
   )
 
   return new PersistentConfigStore(settingsStore, stateStore)
-}
-
-export const mergeConfig = (
-  baseConfig: Record<string, unknown>,
-  overrideConfig: Record<string, unknown>,
-): Record<string, unknown> => {
-  const mergedConfig = { ...baseConfig }
-
-  for (const [key, value] of Object.entries(overrideConfig)) {
-    const baseValue = mergedConfig[key]
-    mergedConfig[key] =
-      isRecord(baseValue) && isRecord(value)
-        ? mergeConfig(baseValue, value)
-        : value
-  }
-
-  return mergedConfig
-}
-
-export const splitConfig = (
-  config: Record<string, unknown>,
-): {
-  settings: Record<string, unknown>
-  state: Record<string, unknown>
-} => {
-  const settings: Record<string, unknown> = {}
-  const state: Record<string, unknown> = {}
-
-  for (const [key, value] of Object.entries(config)) {
-    if (isStateKey(key)) {
-      state[key] = value
-      continue
-    }
-
-    if (isRecord(value)) {
-      const splitValue = splitConfig(value)
-      if (Object.keys(splitValue.settings).length > 0) {
-        settings[key] = splitValue.settings
-      }
-
-      if (Object.keys(splitValue.state).length > 0) {
-        state[key] = splitValue.state
-      }
-
-      continue
-    }
-
-    settings[key] = value
-  }
-
-  return { settings, state }
 }
 
 export class PersistentConfigStore implements ConfigStore {
@@ -178,14 +126,6 @@ const getConfigPath = (): string => {
 
 const getStateConfigPath = (): string => {
   return join(paths.log, configFileName)
-}
-
-const isStateKey = (key: string): boolean => {
-  return (
-    key === currentWorkspaceIdKey ||
-    key === patKey ||
-    key.endsWith(`.${patKey}`)
-  )
 }
 
 const isRecord = (value: unknown): value is Record<string, unknown> => {
